requirements.txt
flask
flask-sockets
gevent
time
datetime
socket
jsonschema
serial
logging
uuid
time
random
websocket_client